src/requirements.txt
asgiref
cryptography
Flask
flask-limiter
gunicorn
lxml
mwclient
mwoauth
pymysql
python-dotenv
requests
tqdm
uvicorn
wikitextparser
CopySVGTranslation>=0.1.6
